Criada uma função para obter um percurso a partir de um cursor e teste para verificar se é possivel ler um percurso a partir da base de dados

+++ app/src/main/java/pt/ipg/provas/Provas.kt
@@ -1,6 +1,8 @@
 package pt.ipg.provas
 
 import android.content.ContentValues
+import android.database.AbstractCursor
+import android.database.Cursor
 import java.util.Date
 
 data class Provas(
@@ -8,7 +10,7 @@ data class Provas(
     var localidade: String,
     var tipo: String,
     var data: String,
-    var id_Percursos: Int,
+    var id_Percursos: Long,
     var id: Long = -1
 ) {
     fun toContentValues(): ContentValues {
